requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.28.1
datasets>=2.10.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
scikit-learn>=1.0.0
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0

# AWS dependencies
boto3>=1.26.0
sagemaker>=2.130.0
awscli>=1.27.0

# Data processing
beautifulsoup4>=4.11.0
lxml>=4.9.0
nltk>=3.7
spacy>=3.5.0
tqdm>=4.64.0
jsonlines>=3.1.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0
evaluate>=0.4.0

# Utilities
PyYAML>=6.0
tensorboard>=2.12.0
wandb>=0.15.0
jupyter>=1.0.0
notebook>=6.5.0
